requirements.txt
requests==2.32.0
pyyaml==6.0
paramiko==3.5.0
docker==6.1.3
jinja2==3.1.2
jsonrpclib
json
